% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_compactor.R
\name{usa_compactor}
\alias{usa_compactor}
\title{Create compact view of the contiguous states of the USA with an inset of Alasaka and/or Hawaii.}
\usage{
usa_compactor(admin_level = 1, scale = "1:50", include_alaska = TRUE,
  include_hawaii = TRUE, save_output = FALSE, output_dir = NULL,
  output_filename = NULL)
}
\arguments{
\item{admin_level}{Numeric value (\emph{0} or \emph{1}) to specify the map administrative level. Default value is \emph{1} (states). If you require a country-level map, change value to \emph{0}.}

\item{scale}{Character string (\emph{1:50} or \emph{1:110}) to specify the scale level of the map. Default value is \emph{1:50}. If you require a higher scale, change value to \emph{1:110}.}

\item{include_alaska}{Logical. Default value is \emph{TRUE}.}

\item{include_hawaii}{Logical. Default value is \emph{TRUE}.}

\item{save_output}{Logical. Should the object be saved to disk as a geojson? Default value is \emph{FALSE}. If set to \emph{TRUE}, and parameters \emph{'output_dir'} and \emph{'output_filename'} are not specified, the file will be saved in the current working directory under the file name: \emph{usa_compact.geojson}.}

\item{output_dir}{Optional character string specifying the path to the directory the geojson should be saved if parameter \emph{'save_file'} is set as \emph{TRUE}. Default value is \emph{NULL}.}

\item{output_filename}{Optional character string specifying the filename under which the geojson should be saved if parameter \emph{'save_file'} is set to \emph{TRUE}. Default value is \emph{NULL}.}
}
\description{
A convenience wrapper building on functions \code{\link[rgdal]{readOGR}},
\code{\link[sp]{spTransform}}, \code{\link[rgeos]{gSimplify}},
\code{\link[geojsonio]{geojson_json}}, and \code{\link[geojsonio]{geojson_write}} to convert a level 0 or 1 level administrative map shapefile from Natural Earth to a geoJSON files of the contiguous states of the USA with an inset of Alaska and/or Hawaii.
}
